requirements.txt
langchain==0.3.0
langchain-community==0.3.0
chromadb==0.5.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
groq==0.9.0
httpx==0.27.2
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
pypdf==4.3.1
python-dotenv==1.0.1
mlflow==2.16.0
torch==2.4.0
python-multipart==0.0.9
